Move Error out of CLI module.
