+++ lib/pivotal_markdown/cli.rb
@@ -1,7 +1,6 @@
 require "thor"
 require "tracker_api"
 
-require_relative "cli/error"
 require_relative "cli/project"
 require_relative "cli/token"
 require_relative "cli/upload"
