+++ todo.md
@@ -1,5 +1,4 @@
 # TODO
-- Move Error out of CLI module.
 - Refactor Verifier.
 - Clear method for default project.
 - Import/adapt upload code.
